PES2UG22CS385_PES2UG22CS413_PES2UG23CS819_PES2UG23CS821_CC-main/requirements.txt
flask>=2.0.0
flask-socketio>=5.3.0
flask-cors>=3.0.10
requests>=2.28.0
docker>=6.0.0
argparse>=1.4.0
supabase>=1.0.3
python-dotenv>=1.0.0
